vim: separate vim and nvim undo and swapdir
Originally, the swap- and undo-directories for vim and nvim were the
same directory. This caused some errors when opening the same files
first in nvim and then in vim, or the other way around. To avoid this
issue, the configuration now saves the nvim undo and swap files to one
directory, while saving the vim files into other directories.

The commit also updates the gitignore and the install script to know
about these new directories.